+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Improved stability and UX of the agent plugin
 - Improved UX in the OSC log widget
 - Text in the OSC log widget can now be selected and copied
+- The palette is now an input field directly in the header
 ### Fixed
 - An issue where OSC play current main selection stopped working
 - The OSC log widget can now scroll to show the full last row